+++ src/toolbox/tools/meta-tools.ts
@@ -11,7 +11,7 @@ function findVersion(): string {
 const version = findVersion();
 
 /**
- * Base directory for the kestrel-cli
+ * Base directory for the skulljs-cli
  */
 const directory = path.join(__dirname, '../..');