+++ src/types/commands/new.d.ts
@@ -1,28 +1,15 @@
 /**
- * Props needed to create a new kestre-cli project
+ * Props needed to create a new slulljs project
  */
 export interface NewProps {
-  name: string
-  backend_path: string | null
-  frontend_path: string | null
-  bname?: string | null
-  fname?: string | null
-  kc_version: string
-}
-
-/**
- * Files needed to be converted to an ejs template from the kestrel-template
- */
-export interface FilesToTransform {
-  toTransform: Array<{
-    path: string
-    part: string
-    filename: string
-    convert: boolean
-    createPath: string
-    replace?: Array<{
-      this: string
-      by: string
-    }>
-  }>
+  name: string;
+  backend_path: string | null;
+  backend_repository: string | null;
+  backend_version: string | null;
+  frontend_path: string | null;
+  frontend_repository: string | null;
+  frontend_version: string | null;
+  bname?: string | null;
+  fname?: string | null;
+  sk_version: string;
 }