+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.9.2 - sk new
+
+- Add sk new
+
+## v0.9.1 - Toolbox + Structure
+
+- Add toolbox + structure
+
 ## v0.9.0 - Init
 
 - Init of the project